Renamed opToAssembly.ts to createInstruction.ts
