+++ dev/src/SmartC/codeGenerator/astProcessor/endAsn.ts
@@ -1,6 +1,6 @@
 import { CONTRACT } from '../../typings/contractTypes'
 import { END_ASN } from '../../typings/syntaxTypes'
-import { createSimpleInstruction, createInstruction } from '../assemblyProcessor/opToAssembly'
+import { createSimpleInstruction, createInstruction } from '../assemblyProcessor/createInstruction'
 import { GENCODE_AUXVARS, GENCODE_ARGS, GENCODE_SOLVED_OBJECT } from '../typings/codeGeneratorTypes'
 import { utils } from '../utils'
 import { genCode } from './genCode'
